panel: hk3: allow overriding frequency for HBM mode
Store commands for frequency control as a struct, and make a copy
of the struct and values inside for HBM. These copies can also be
modified via kernel parameters with _hbm in their names.

Default values are unchanged. This commit also aims to provide further
possibility, same as last commit.